fix cast error on non-Mac platforms and clarify comment

+++ display_list/display_list_vertices.cc
@@ -160,7 +160,7 @@ DlVertices::DlVertices(DlVertexMode mode,
       offset += bytes;
       return ret;
     } else {
-      return 0ul;
+      return size_t(0);
     }
   };
